move almost everything out of windows module
