+++ rust/headless-client/src/linux.rs
@@ -189,6 +189,28 @@ pub fn firezone_group() -> Result<nix::unistd::Group> {
     Ok(group)
 }
 
+pub(crate) struct IpcServer {}
+
+impl IpcServer {
+    /// Platform-specific setup
+    pub(crate) async fn new() -> Result<()> {
+        wintun_install::ensure_dll()?;
+        Ok(())
+    }
+
+    pub(crate) async fn next_client(&mut self) -> Result<IpcStream> {
+        let server = create_pipe_server()?;
+        tracing::info!("Listening for GUI to connect over IPC...");
+        server
+            .connect()
+            .await
+            .context("Couldn't accept IPC connection from GUI")?;
+        Ok(server)
+    }
+}
+
+pub(crate) struct IpcStream(named_pipe::NamedPipeServer);
+
 pub(crate) async fn ipc_listen(cli: CliCommon) -> Result<()> {
     // Remove the socket if a previous run left it there
     let sock_path = sock_path();
@@ -217,102 +239,6 @@ pub(crate) async fn ipc_listen(cli: CliCommon) -> Result<()> {
             tracing::error!(?error, "Error while handling IPC client");
         }
     }
-}
-
-#[derive(Clone)]
-struct CallbackHandlerIpc {
-    cb_tx: mpsc::Sender<IpcServerMsg>,
-}
-
-impl Callbacks for CallbackHandlerIpc {
-    fn on_disconnect(&self, error: &connlib_client_shared::Error) {
-        tracing::error!(?error, "Got `on_disconnect` from connlib");
-        self.cb_tx
-            .try_send(IpcServerMsg::OnDisconnect)
-            .expect("should be able to send OnDisconnect");
-    }
-
-    fn on_set_interface_config(
-        &self,
-        ipv4: Ipv4Addr,
-        ipv6: Ipv6Addr,
-        dns: Vec<IpAddr>,
-    ) -> Option<i32> {
-        tracing::info!("TunnelReady (on_set_interface_config)");
-        self.cb_tx
-            .try_send(IpcServerMsg::OnSetInterfaceConfig { ipv4, ipv6, dns })
-            .expect("Should be able to send TunnelReady");
-        None
-    }
-
-    fn on_update_resources(&self, resources: Vec<callbacks::ResourceDescription>) {
-        tracing::debug!(len = resources.len(), "New resource list");
-        self.cb_tx
-            .try_send(IpcServerMsg::OnUpdateResources(resources))
-            .expect("Should be able to send OnUpdateResources");
-    }
-}
-
-async fn handle_ipc_client(cli: &CliCommon, stream: UnixStream) -> Result<()> {
-    let (rx, tx) = stream.into_split();
-    let mut rx = FramedRead::new(rx, LengthDelimitedCodec::new());
-    let mut tx = FramedWrite::new(tx, LengthDelimitedCodec::new());
-    let (cb_tx, mut cb_rx) = mpsc::channel(100);
-
-    let send_task = tokio::spawn(async move {
-        while let Some(msg) = cb_rx.recv().await {
-            tx.send(serde_json::to_string(&msg)?.into()).await?;
-        }
-        Ok::<_, anyhow::Error>(())
-    });
-
-    let mut connlib = None;
-    let callback_handler = CallbackHandlerIpc { cb_tx };
-    while let Some(msg) = rx.next().await {
-        let msg = msg?;
-        let msg: super::IpcClientMsg = serde_json::from_slice(&msg)?;
-
-        match msg {
-            IpcClientMsg::Connect { api_url, token } => {
-                let token = secrecy::SecretString::from(token);
-                assert!(connlib.is_none());
-                let device_id = connlib_shared::device_id::get()
-                    .context("Failed to read / create device ID")?;
-                let (private_key, public_key) = keypair();
-
-                let login = LoginUrl::client(
-                    Url::parse(&api_url)?,
-                    &token,
-                    device_id.id,
-                    None,
-                    public_key.to_bytes(),
-                )?;
-
-                connlib = Some(connlib_client_shared::Session::connect(
-                    login,
-                    Sockets::new(),
-                    private_key,
-                    None,
-                    callback_handler.clone(),
-                    cli.max_partition_time
-                        .map(|t| t.into())
-                        .or(Some(std::time::Duration::from_secs(60 * 60 * 24 * 30))),
-                    tokio::runtime::Handle::try_current()?,
-                ));
-            }
-            IpcClientMsg::Disconnect => {
-                if let Some(connlib) = connlib.take() {
-                    connlib.disconnect();
-                }
-            }
-            IpcClientMsg::Reconnect => connlib.as_mut().context("No connlib session")?.reconnect(),
-            IpcClientMsg::SetDns(v) => connlib.as_mut().context("No connlib session")?.set_dns(v),
-        }
-    }
-
-    send_task.abort();
-
-    Ok(())
 }
 
 /// Platform-specific setup needed for connlib
